Remove 'busy' state
Honestly can't remember what the intent behind it was, there may be a
race somewhere that it was trying to avoid. But as written it broke
removing breakpoints as F5ing would set the state to 'busy' and then
the debugger would lose control back to Unreal. This prevented setting
any new breakpoints as the set breakpoints handler would block until
the state was not busy.

May need to re-add something like this back in later if it was working
around something else.