Fix crashes from bad encodings
Strings coming from Unreal with game data (e.g. variable string content)
are encoded with the game's encoding. For INT this is iso8859-1, and may
include accented characters. The JSON encoding used by cppdap wants
everything to be utf-8 encoded, and some of these characters are not
valid utf-8 and will throw exceptions.

This is a quick fix for some languages only - strings coming from the game
are converted from iso8859-1 to utf-8 before being sent to cppdap.
Languages other than western european ones use UTF-16 internally, so
these can also be relatively easily converted to UTF-8 and back, but we
need to know what language the game is running in and I don't know how
to detect that right now. Alternatively the language can be added to the
adapter command line.